Fix tests and update intl version
