+++ pubspec.yaml
@@ -44,7 +44,7 @@ dependencies:
   # flutter_native_timezone: ^2.0.0
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  intl: ^0.19.0
+  intl: ^0.20.2
   awesome_notifications_core: ^0.9.3
   awesome_notifications: any
   firebase_core: ^3.1.1
